backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.73",
    "fastapi>=0.121.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
